Add files via upload
Esav was "ish sade", he went to the field, where he found the Yid'oni (some impure, future-teller creature), who told him the future and then he told it to his father, Yitzchok, making him believe that he has ruach hakodesh. Yaakov did not do that, he was an "ish tam". The prohibition of not to turn to the Yid'oni followed by "tamim tihye". (Devarim 18:11-13). The pasuk in our parasha (Bereshis 25:27) refers to this, when it calls Yaakov an "ish tam". (Vilna Gaon and Chasam Sofer)